refactor: images carousel
